Add SensitiveParameter and some Credentials tests
